Rearrange old unittests.py; Use nosetests
* Move unittests.py to tests/ subdirectory
* Split out dependencyparser test to an extra file
* Add configuration for nosetests
* Use nosetests in Makefile target "check"
