+++ nosetests.cfg
@@ -1,0 +1,14 @@
+[nosetests]
+# create test coverage report
+with-coverage=1
+with-coverage=1
+cover-html=1
+cover-html-dir=build/html
+cover-package=piupartslib,piuparts
+#cover-inclusive=1
+
+# run doctests
+with-doctest=1
+
+# catch all tests
+# all-modules=1
